computed discord id

+++ Records/DiscordLink.cs
@@ -3,5 +3,5 @@ namespace CmApi.Records;
 public record DiscordLink(
     int Id,
     int PlayerId,
-    int DiscordId
+    long DiscordId
     );
